Update README.md
Since the start of the COVID-19 pandemic, researchers have been striving to help society face many problems caused by this pandemic. Among the recent initiatives, one has drawn the authors’ attention: producing low-cost, open-source mechanical ventilators. The motivation comes from the worldwide shortage of mechanical ventilators in the treatment of COVID-19 patients—mechanical ventilators keep severely ill patients alive. This paper contributes to this initiative. This has detailed the construction of a functional, low-cost, and open-source mechanical ventilator. The authors’ contribution to this topic aims to mitigate the effects of this worldwide ventilator’s shortage—a shocking, unfortunate event that hits hard deprived areas. This has shown a numerical method that can monitor, in real-time, whether the patient has a healthy or unhealthy pulmonary condition. This useful yet straightforward numerical method opens up the possibility of applications in other mechanical ventilators as well. In summary, this paper contributes to both fronts—theory and practice. Alarms can be included in this project, using either an alarm screen or speakers, like the ones that alert clinicians when the pressure reaches some threshold values.